Revert "feat: deduplicate generation requests (#13299)"
This reverts commit 599bb86c3558b27feccd31c857b8a4ac516e9cbd.
