+++ gen.pollinations.ai/src/middleware/auth.ts
@@ -7,7 +7,6 @@ import {
     StagingAccessDeniedError,
 } from "@shared/auth/api-key.ts";
 import type { CommunityEndpointRuntime } from "@shared/community-endpoints.ts";
-import type { Context } from "hono";
 import { createMiddleware } from "hono/factory";
 import { HTTPException } from "hono/http-exception";
 import type { LoggerVariables } from "./logger.ts";
@@ -31,101 +30,75 @@ export type AuthVariables = {
     };
 };
 
-export type GenerationAuthSnapshot = {
-    user: Pick<AuthUser, "id" | "tier">;
-    apiKey?: Omit<AuthenticatedApiKey, "rawKey">;
-    agentRun?: AgentRunClaims;
-};
-
 export type AuthEnv = {
     Bindings: CloudflareBindings;
     Variables: LoggerVariables & AuthVariables & Partial<ModelVariables>;
 };
 
-function installAuth(
-    c: Context<AuthEnv>,
-    authResult: {
-        user?: AuthUser;
-        apiKey?: AuthenticatedApiKey;
-        agentRun?: AgentRunClaims;
-    },
-): void {
-    const { user, apiKey, agentRun } = authResult;
-
-    const requireAuthorization = async (options?: {
-        message?: string;
-    }): Promise<void> => {
-        if (!user) {
-            throw new HTTPException(401, {
-                message: options?.message,
-            });
-        }
-    };
-
-    const requireUser = (): AuthUser => {
-        if (!user) throw new HTTPException(401);
-        return user;
-    };
-
-    function requireModelAccess(): void {
-        const model = c.var.model;
-        if (!model) return;
-
-        if (agentRun && model.communityEndpoint) {
-            throw new HTTPException(403, {
-                message: "Agent run tokens cannot call community models",
-            });
-        }
-
-        if (!apiKey?.permissions?.models) return;
-
-        if (!apiKey.permissions.models.includes(model.resolved)) {
-            throw new HTTPException(403, {
-                message: `Model '${model.requested}' is not allowed for this API key`,
-            });
-        }
-    }
-
-    c.set("auth", {
-        user,
-        apiKey,
-        requireAuthorization,
-        requireUser,
-        requireModelAccess,
-        ...(agentRun && { agentRun }),
-    });
-}
-
 export const auth = () =>
     createMiddleware<AuthEnv>(async (c, next) => {
-        let authResult: Awaited<ReturnType<typeof authenticateApiKeyRequest>>;
-        try {
-            authResult = await authenticateApiKeyRequest({
-                request: c.req.raw,
-                env: c.env,
-                ctx: c.executionCtx,
-            });
-        } catch (error) {
-            if (
-                error instanceof BannedAccountError ||
-                error instanceof StagingAccessDeniedError
-            ) {
-                throw new HTTPException(403, {
-                    message: error.message,
+        const authResult = await (async () => {
+            try {
+                return await authenticateApiKeyRequest({
+                    request: c.req.raw,
+                    env: c.env,
+                    ctx: c.executionCtx,
+                });
+            } catch (error) {
+                if (
+                    error instanceof BannedAccountError ||
+                    error instanceof StagingAccessDeniedError
+                ) {
+                    throw new HTTPException(403, { message: error.message });
+                }
+                throw error;
+            }
+        })();
+
+        const { user, apiKey, agentRun } = authResult || {};
+
+        const requireAuthorization = async (options?: {
+            message?: string;
+        }): Promise<void> => {
+            if (!user) {
+                throw new HTTPException(401, {
+                    message: options?.message,
                 });
             }
-            throw error;
-        }
-        installAuth(c, authResult || {});
-        await next();
-    });
+        };
 
-export const authFromSnapshot = (snapshot: GenerationAuthSnapshot) =>
-    createMiddleware<AuthEnv>(async (c, next) => {
-        installAuth(c, {
-            user: snapshot.user as AuthUser,
-            apiKey: snapshot.apiKey as AuthenticatedApiKey | undefined,
-            agentRun: snapshot.agentRun,
+        const requireUser = (): AuthUser => {
+            if (!user) throw new HTTPException(401);
+            return user;
+        };
+
+        function requireModelAccess(): void {
+            const model = c.var.model;
+            if (!model) return;
+
+            if (agentRun && model.communityEndpoint) {
+                throw new HTTPException(403, {
+                    message: "Agent run tokens cannot call community models",
+                });
+            }
+
+            if (!apiKey?.permissions?.models) return;
+
+            if (!apiKey.permissions.models.includes(model.resolved)) {
+                throw new HTTPException(403, {
+                    message: `Model '${model.requested}' is not allowed for this API key`,
+                });
+            }
+        }
+
+        c.set("auth", {
+            user,
+            apiKey,
+            requireAuthorization,
+            requireUser,
+            requireModelAccess,
+            ...(agentRun && { agentRun }),
         });
+
         await next();
     });
