+++ gen.pollinations.ai/src/routes/audio.ts
@@ -22,9 +22,6 @@ import { z } from "zod";
 import type { Env } from "@/env.ts";
 import { auth } from "@/middleware/auth.ts";
 import { balance } from "@/middleware/balance.ts";
-import { prepareGenerationRequest } from "@/middleware/generation-cache.ts";
-import { deduplicateGeneration } from "@/middleware/generation-deduplication.ts";
-import { audioCache } from "@/middleware/media-cache.ts";
 import { resolveModel } from "@/middleware/model.ts";
 import { frontendKeyRateLimit } from "@/middleware/rate-limit-durable.ts";
 import { edgeRateLimit } from "@/middleware/rate-limit-edge.ts";
@@ -33,11 +30,10 @@ import {
     applySafetyToTexts,
     withSafetyHeaders,
 } from "@/middleware/safety.ts";
-import { textCache } from "@/middleware/text-cache.ts";
 import { track } from "@/middleware/track.ts";
 import googleCloudAuth from "@/text/auth/googleCloudAuth.ts";
-import { arrayBufferToBase64, normalizeSeed } from "@/util.ts";
-import { generationAccess } from "@/utils/generation-access.ts";
+import { arrayBufferToBase64 } from "@/util.ts";
+import { requireGenerationAccess } from "@/utils/generation-access.ts";
 import {
     type FallbackCandidate,
     withModelFallbackResponse,
@@ -126,7 +122,7 @@ const CreateSpeechRequestSchema = z
             description:
                 "ElevenLabs composition_plan for music generation/inpainting. Cannot be combined with a plain prompt upstream.",
         }),
-        seed: z.number().int().min(0).max(4294967295).optional().meta({
+        seed: z.number().int().min(-1).max(4294967295).optional().meta({
             description:
                 "Seed for deterministic output. Same seed + params = best-effort return of the same cached result. Omit for random.",
             example: 42,
@@ -155,7 +151,7 @@ const CreateDialogueRequestSchema = z
         response_format: z
             .enum(["mp3", "opus", "aac", "wav", "pcm"])
             .default("mp3"),
-        seed: z.number().int().min(0).max(4294967295).optional(),
+        seed: z.number().int().min(-1).max(4294967295).optional(),
         safe: SafeSchema,
     })
     .refine(
@@ -2515,7 +2511,7 @@ export async function handleSimpleAudio(c: AudioContext): Promise<Response> {
             text,
             voice: query.voice,
             responseFormat: query.response_format,
-            seed: normalizeSeed(query.seed),
+            seed: query.seed,
             duration: query.duration,
             seconds: query.seconds,
             steps: query.steps,
@@ -2532,310 +2528,6 @@ export async function handleSimpleAudio(c: AudioContext): Promise<Response> {
             log,
         }),
     );
-}
-
-export async function handleDialogue(c: AudioContext): Promise<Response> {
-    const log = c.get("log").getChild("dialogue");
-    let request: z.infer<typeof CreateDialogueRequestSchema>;
-    try {
-        request = CreateDialogueRequestSchema.parse(await c.req.json());
-    } catch (error) {
-        if (error instanceof z.ZodError) {
-            throw new UpstreamError(400 as ContentfulStatusCode, {
-                message: error.issues[0]?.message || "Invalid request body",
-            });
-        }
-        throw new UpstreamError(400 as ContentfulStatusCode, {
-            message: "Invalid JSON body",
-        });
-    }
-
-    const safeTexts = await applySafetyToTexts(
-        c,
-        request.inputs.map((input) => input.text),
-        request.safe,
-    );
-    const inputs = request.inputs.map((input, index) => ({
-        ...input,
-        text: safeTexts[index],
-    }));
-    const response = await generateElevenLabsDialogue({
-        inputs,
-        responseFormat: request.response_format,
-        seed: request.seed,
-        apiKey: c.env.ELEVENLABS_API_KEY,
-        log,
-    });
-    return withSafetyHeaders(c, response);
-}
-
-export async function handleVoiceChanger(c: AudioContext): Promise<Response> {
-    const log = c.get("log").getChild("voice-changer");
-    const formData = c.get("formData");
-    if (!formData) {
-        throw new UpstreamError(400 as ContentfulStatusCode, {
-            message: "Invalid multipart form data",
-        });
-    }
-
-    const audio = formData.get("audio");
-    if (!(audio instanceof File)) {
-        throw new UpstreamError(400 as ContentfulStatusCode, {
-            message: "Missing required audio file.",
-        });
-    }
-    const voice = formData.get("voice");
-    const responseFormat = formData.get("response_format");
-    const resolvedFormat =
-        typeof responseFormat === "string" && responseFormat !== ""
-            ? responseFormat
-            : "mp3";
-    if (!["mp3", "opus", "aac", "wav", "pcm"].includes(resolvedFormat)) {
-        throw new UpstreamError(400 as ContentfulStatusCode, {
-            message: "response_format must be mp3, opus, aac, wav, or pcm.",
-        });
-    }
-
-    return changeVoiceWithElevenLabs({
-        audio,
-        voice: typeof voice === "string" && voice !== "" ? voice : "alloy",
-        responseFormat: resolvedFormat,
-        apiKey: c.env.ELEVENLABS_API_KEY,
-        log,
-    });
-}
-
-export async function handleVoiceIsolator(c: AudioContext): Promise<Response> {
-    const log = c.get("log").getChild("voice-isolator");
-    const formData = c.get("formData");
-    if (!formData) {
-        throw new UpstreamError(400 as ContentfulStatusCode, {
-            message: "Invalid multipart form data",
-        });
-    }
-    const audio = formData.get("audio");
-    if (!(audio instanceof File)) {
-        throw new UpstreamError(400 as ContentfulStatusCode, {
-            message: "Missing required audio file.",
-        });
-    }
-    return isolateVoiceWithElevenLabs({
-        audio,
-        apiKey: c.env.ELEVENLABS_API_KEY,
-        log,
-    });
-}
-
-export async function handleSpeech(c: AudioContext): Promise<Response> {
-    const log = c.get("log").getChild("tts");
-    const {
-        input,
-        safe,
-        voice,
-        response_format,
-        duration,
-        seconds,
-        steps,
-        negative_prompt,
-        instrumental,
-        store_for_inpainting,
-        conditioning_ref,
-        composition_plan,
-        reference_audio,
-        seed,
-        instruct,
-        loop,
-        prompt_influence,
-    } = await parseSpeechRequest(c);
-    requireElevenMusicOptions(c.var.model.resolved, {
-        referenceAudio: reference_audio,
-        compositionPlan: composition_plan,
-        conditioningRef: conditioning_ref,
-        storeForInpainting: store_for_inpainting,
-    });
-    const safeInput = await applySafety(c, input, safe);
-    const referenceAudio = reference_audio
-        ? await fetchReferenceAudio(reference_audio)
-        : undefined;
-
-    return withAudioFallback(c, (candidate) =>
-        dispatchAudioGeneration(c, candidate.id, {
-            text: safeInput,
-            voice,
-            responseFormat: response_format,
-            seed,
-            duration,
-            seconds,
-            steps,
-            negativePrompt: negative_prompt,
-            instrumental,
-            storeForInpainting: store_for_inpainting,
-            conditioningRef: conditioning_ref,
-            compositionPlan: composition_plan,
-            referenceAudio,
-            instruct,
-            loop,
-            promptInfluence: prompt_influence,
-            apiKey: c.env.ELEVENLABS_API_KEY,
-            dashScopeApiKey: c.env.DASHSCOPE_API_KEY,
-            deepInfraApiKey: c.env.DEEPINFRA_API_KEY,
-            falKey: c.env.FAL_KEY,
-            stabilityApiKey: c.env.STABILITY_API_KEY,
-            log,
-        }),
-    );
-}
-
-export async function handleSpeechWithTimestamps(
-    c: AudioContext,
-): Promise<Response> {
-    const log = c.get("log").getChild("tts-timestamps");
-    const { input, safe, voice, response_format, seed } =
-        await parseSpeechRequest(c);
-    const modelName = c.var.model.resolved;
-    if (!(modelName in ELEVENLABS_TTS_MODEL_IDS)) {
-        throw new UpstreamError(400 as ContentfulStatusCode, {
-            message:
-                "Timestamped speech supports elevenlabs, elevenflash, and eleven-multilingual-v2.",
-        });
-    }
-    if (response_format === "flac") {
-        throw new UpstreamError(400 as ContentfulStatusCode, {
-            message:
-                "Timestamped speech supports mp3, opus, aac, wav, and pcm output.",
-        });
-    }
-    const safeInput = await applySafety(c, input, safe);
-    return withAudioFallback(c, (candidate) =>
-        generateElevenLabsSpeechWithTimestamps({
-            modelName: candidate.id as ElevenLabsTtsModelName,
-            text: safeInput,
-            voice,
-            responseFormat: response_format,
-            seed,
-            apiKey: c.env.ELEVENLABS_API_KEY,
-            log,
-        }),
-    );
-}
-
-export async function handleTranscription(c: AudioContext): Promise<Response> {
-    const log = c.get("log").getChild("transcription");
-    const formData = c.get("formData");
-    if (!formData) {
-        throw new UpstreamError(400 as ContentfulStatusCode, {
-            message: "Invalid multipart form data",
-        });
-    }
-
-    const file = formData.get("file") as File;
-    const language = formData.get("language") as string | null;
-    const prompt = formData.get("prompt") as string | null;
-    const responseFormat = formData.get("response_format") as string | null;
-    const temperatureRaw = formData.get("temperature") as string | null;
-    const temperature =
-        temperatureRaw !== null && temperatureRaw !== ""
-            ? Number(temperatureRaw)
-            : undefined;
-    const speakersExpected = parsePositiveInt(
-        formData.get("speakers_expected"),
-        "speakers_expected",
-    );
-
-    if (!file) {
-        throw new UpstreamError(400 as ContentfulStatusCode, {
-            message: "Missing required field: file",
-        });
-    }
-    if (speakersExpected !== undefined && responseFormat !== "diarized_json") {
-        throw new UpstreamError(400 as ContentfulStatusCode, {
-            message: "speakers_expected requires response_format=diarized_json",
-        });
-    }
-
-    const result = await withAudioFallback(c, async (candidate) => {
-        if (candidate.id === "grok-transcribe") {
-            return transcribeWithXai({
-                file,
-                language: language || undefined,
-                responseFormat: responseFormat || undefined,
-                apiKey: c.env.XAI_API_KEY,
-                log,
-            });
-        }
-        if (candidate.id === "scribe") {
-            return transcribeWithElevenLabs({
-                file,
-                language: language || undefined,
-                responseFormat: responseFormat || undefined,
-                apiKey: c.env.ELEVENLABS_API_KEY,
-                log,
-                numSpeakers: speakersExpected,
-            });
-        }
-        if (
-            candidate.id === "universal-2" ||
-            candidate.id === "universal-3.5-pro"
-        ) {
-            return transcribeWithAssemblyAi({
-                file,
-                language: language || undefined,
-                prompt: prompt || undefined,
-                responseFormat: responseFormat || undefined,
-                temperature,
-                model: candidate.id,
-                apiKey: (c.env as unknown as { ASSEMBLYAI_API_KEY: string })
-                    .ASSEMBLYAI_API_KEY,
-                log,
-                speakersExpected,
-            });
-        }
-
-        const ovhApiKey = c.env.OVHCLOUD_API_KEY;
-        if (!ovhApiKey) {
-            throw new UpstreamError(500 as ContentfulStatusCode, {
-                message:
-                    "Transcription service is not configured (missing API key)",
-            });
-        }
-        validateWhisperResponseFormat(responseFormat);
-
-        const whisperFormData = new FormData();
-        const filename =
-            file.name && file.name !== "blob" ? file.name : "audio.mp3";
-        whisperFormData.append("file", file, filename);
-        if (language) whisperFormData.append("language", language);
-        whisperFormData.append("response_format", "verbose_json");
-        whisperFormData.append("model", "whisper-large-v3");
-        whisperFormData.append("timestamp_granularities[]", "word");
-
-        const whisperUrl =
-            "https://oai.endpoints.kepler.ai.cloud.ovh.net/v1/audio/transcriptions";
-        const response = await ensureUpstreamOk(
-            await fetch(whisperUrl, {
-                method: "POST",
-                headers: { Authorization: `Bearer ${ovhApiKey}` },
-                body: whisperFormData,
-            }),
-            whisperUrl,
-        );
-
-        let whisper: WhisperVerboseJson;
-        try {
-            whisper = JSON.parse(await response.text());
-        } catch {
-            throw new UpstreamError(502 as ContentfulStatusCode, {
-                message: "Whisper returned an unexpected (non-JSON) response",
-            });
-        }
-        const usageHeaders = buildUsageHeaders(
-            candidate.id,
-            createAudioSecondsUsage(extractWhisperUsage(whisper, log)),
-        );
-        return formatWhisperResponse(whisper, responseFormat, usageHeaders);
-    });
-    c.var.track.overrideResponseTracking(result.clone());
-    return result;
 }
 
 export const audioRoutes = new Hono<Env>()
@@ -2921,11 +2613,43 @@ export const audioRoutes = new Hono<Env>()
             supportedEndpoint: "/v1/audio/dialogue",
         }),
         track("generate.audio"),
-        prepareGenerationRequest,
-        audioCache,
-        generationAccess,
-        deduplicateGeneration,
-        handleDialogue,
+        async (c) => {
+            const log = c.get("log").getChild("dialogue");
+            await requireGenerationAccess(c.var, c.env);
+
+            let request: z.infer<typeof CreateDialogueRequestSchema>;
+            try {
+                request = CreateDialogueRequestSchema.parse(await c.req.json());
+            } catch (error) {
+                if (error instanceof z.ZodError) {
+                    throw new UpstreamError(400 as ContentfulStatusCode, {
+                        message:
+                            error.issues[0]?.message || "Invalid request body",
+                    });
+                }
+                throw new UpstreamError(400 as ContentfulStatusCode, {
+                    message: "Invalid JSON body",
+                });
+            }
+
+            const safeTexts = await applySafetyToTexts(
+                c,
+                request.inputs.map((input) => input.text),
+                request.safe,
+            );
+            const inputs = request.inputs.map((input, index) => ({
+                ...input,
+                text: safeTexts[index],
+            }));
+            const response = await generateElevenLabsDialogue({
+                inputs,
+                responseFormat: request.response_format,
+                seed: request.seed,
+                apiKey: c.env.ELEVENLABS_API_KEY,
+                log,
+            });
+            return withSafetyHeaders(c, response);
+        },
     )
     .post(
         "/voice-changer",
@@ -2997,11 +2721,49 @@ export const audioRoutes = new Hono<Env>()
             supportedEndpoint: "/v1/audio/voice-changer",
         }),
         track("generate.audio"),
-        prepareGenerationRequest,
-        audioCache,
-        generationAccess,
-        deduplicateGeneration,
-        handleVoiceChanger,
+        async (c) => {
+            const log = c.get("log").getChild("voice-changer");
+            await requireGenerationAccess(c.var, c.env);
+
+            let formData: FormData;
+            try {
+                formData = c.get("formData") || (await c.req.formData());
+            } catch {
+                throw new UpstreamError(400 as ContentfulStatusCode, {
+                    message: "Invalid multipart form data",
+                });
+            }
+
+            const audio = formData.get("audio");
+            if (!(audio instanceof File)) {
+                throw new UpstreamError(400 as ContentfulStatusCode, {
+                    message: "Missing required audio file.",
+                });
+            }
+            const voice = formData.get("voice");
+            const responseFormat = formData.get("response_format");
+            const resolvedFormat =
+                typeof responseFormat === "string" && responseFormat !== ""
+                    ? responseFormat
+                    : "mp3";
+            if (
+                !["mp3", "opus", "aac", "wav", "pcm"].includes(resolvedFormat)
+            ) {
+                throw new UpstreamError(400 as ContentfulStatusCode, {
+                    message:
+                        "response_format must be mp3, opus, aac, wav, or pcm.",
+                });
+            }
+
+            return changeVoiceWithElevenLabs({
+                audio,
+                voice:
+                    typeof voice === "string" && voice !== "" ? voice : "alloy",
+                responseFormat: resolvedFormat,
+                apiKey: c.env.ELEVENLABS_API_KEY,
+                log,
+            });
+        },
     )
     .post(
         "/voice-isolator",
@@ -3051,11 +2813,32 @@ export const audioRoutes = new Hono<Env>()
             supportedEndpoint: "/v1/audio/voice-isolator",
         }),
         track("generate.audio"),
-        prepareGenerationRequest,
-        audioCache,
-        generationAccess,
-        deduplicateGeneration,
-        handleVoiceIsolator,
+        async (c) => {
+            const log = c.get("log").getChild("voice-isolator");
+            await requireGenerationAccess(c.var, c.env);
+
+            let formData: FormData;
+            try {
+                formData = c.get("formData") || (await c.req.formData());
+            } catch {
+                throw new UpstreamError(400 as ContentfulStatusCode, {
+                    message: "Invalid multipart form data",
+                });
+            }
+
+            const audio = formData.get("audio");
+            if (!(audio instanceof File)) {
+                throw new UpstreamError(400 as ContentfulStatusCode, {
+                    message: "Missing required audio file.",
+                });
+            }
+
+            return isolateVoiceWithElevenLabs({
+                audio,
+                apiKey: c.env.ELEVENLABS_API_KEY,
+                log,
+            });
+        },
     )
     .post(
         "/speech",
@@ -3180,11 +2963,70 @@ export const audioRoutes = new Hono<Env>()
             supportedEndpoint: "/v1/audio/speech",
         }),
         track("generate.audio"),
-        prepareGenerationRequest,
-        audioCache,
-        generationAccess,
-        deduplicateGeneration,
-        handleSpeech,
+        async (c) => {
+            const log = c.get("log").getChild("tts");
+            await requireGenerationAccess(c.var, c.env);
+
+            const {
+                input,
+                safe,
+                voice,
+                response_format,
+                duration,
+                seconds,
+                steps,
+                negative_prompt,
+                instrumental,
+                store_for_inpainting,
+                conditioning_ref,
+                composition_plan,
+                reference_audio,
+                seed,
+                instruct,
+                loop,
+                prompt_influence,
+            } = await parseSpeechRequest(c);
+            requireElevenMusicOptions(c.var.model.resolved, {
+                referenceAudio: reference_audio,
+                compositionPlan: composition_plan,
+                conditioningRef: conditioning_ref,
+                storeForInpainting: store_for_inpainting,
+            });
+            const safeInput = await applySafety(c, input, safe);
+            const referenceAudio = reference_audio
+                ? await fetchReferenceAudio(reference_audio)
+                : undefined;
+
+            const apiKey = (c.env as unknown as { ELEVENLABS_API_KEY: string })
+                .ELEVENLABS_API_KEY;
+
+            return withAudioFallback(c, (candidate) =>
+                dispatchAudioGeneration(c, candidate.id, {
+                    text: safeInput,
+                    voice,
+                    responseFormat: response_format,
+                    seed,
+                    duration,
+                    seconds,
+                    steps,
+                    negativePrompt: negative_prompt,
+                    instrumental,
+                    storeForInpainting: store_for_inpainting,
+                    conditioningRef: conditioning_ref,
+                    compositionPlan: composition_plan,
+                    referenceAudio,
+                    instruct,
+                    loop,
+                    promptInfluence: prompt_influence,
+                    apiKey,
+                    dashScopeApiKey: c.env.DASHSCOPE_API_KEY,
+                    deepInfraApiKey: c.env.DEEPINFRA_API_KEY,
+                    falKey: c.env.FAL_KEY,
+                    stabilityApiKey: c.env.STABILITY_API_KEY,
+                    log,
+                }),
+            );
+        },
     )
     .post(
         "/speech/with-timestamps",
@@ -3301,11 +3143,39 @@ export const audioRoutes = new Hono<Env>()
             supportedEndpoint: "/v1/audio/speech/with-timestamps",
         }),
         track("generate.audio"),
-        prepareGenerationRequest,
-        textCache,
-        generationAccess,
-        deduplicateGeneration,
-        handleSpeechWithTimestamps,
+        async (c) => {
+            const log = c.get("log").getChild("tts-timestamps");
+            await requireGenerationAccess(c.var, c.env);
+
+            const { input, safe, voice, response_format, seed } =
+                await parseSpeechRequest(c);
+            const modelName = c.var.model.resolved;
+            if (!(modelName in ELEVENLABS_TTS_MODEL_IDS)) {
+                throw new UpstreamError(400 as ContentfulStatusCode, {
+                    message:
+                        "Timestamped speech supports elevenlabs, elevenflash, and eleven-multilingual-v2.",
+                });
+            }
+            if (response_format === "flac") {
+                throw new UpstreamError(400 as ContentfulStatusCode, {
+                    message:
+                        "Timestamped speech supports mp3, opus, aac, wav, and pcm output.",
+                });
+            }
+            const safeInput = await applySafety(c, input, safe);
+
+            return withAudioFallback(c, (candidate) =>
+                generateElevenLabsSpeechWithTimestamps({
+                    modelName: candidate.id as ElevenLabsTtsModelName,
+                    text: safeInput,
+                    voice,
+                    responseFormat: response_format,
+                    seed,
+                    apiKey: c.env.ELEVENLABS_API_KEY,
+                    log,
+                }),
+            );
+        },
     )
     .post(
         "/transcriptions",
@@ -3428,11 +3298,149 @@ export const audioRoutes = new Hono<Env>()
             supportedEndpoint: "/v1/audio/transcriptions",
         }),
         track("generate.audio"),
-        prepareGenerationRequest,
-        textCache,
-        generationAccess,
-        deduplicateGeneration,
-        handleTranscription,
+        async (c) => {
+            const log = c.get("log").getChild("transcription");
+            await requireGenerationAccess(c.var, c.env);
+
+            // Get formData from middleware or parse it
+            let formData: FormData;
+            try {
+                formData = c.get("formData") || (await c.req.formData());
+            } catch (error) {
+                log.warn("Invalid multipart form data: {message}", {
+                    message:
+                        error instanceof Error ? error.message : String(error),
+                });
+                throw new UpstreamError(400 as ContentfulStatusCode, {
+                    message: "Invalid multipart form data",
+                });
+            }
+
+            const file = formData.get("file") as File;
+            const language = formData.get("language") as string | null;
+            const prompt = formData.get("prompt") as string | null;
+            const responseFormat = formData.get("response_format") as
+                | string
+                | null;
+            const temperatureRaw = formData.get("temperature") as string | null;
+            const temperature =
+                temperatureRaw !== null && temperatureRaw !== ""
+                    ? Number(temperatureRaw)
+                    : undefined;
+            const speakersExpected = parsePositiveInt(
+                formData.get("speakers_expected"),
+                "speakers_expected",
+            );
+            const wantsDiarizedJson = responseFormat === "diarized_json";
+
+            if (!file) {
+                throw new UpstreamError(400 as ContentfulStatusCode, {
+                    message: "Missing required field: file",
+                });
+            }
+
+            if (speakersExpected !== undefined && !wantsDiarizedJson) {
+                throw new UpstreamError(400 as ContentfulStatusCode, {
+                    message:
+                        "speakers_expected requires response_format=diarized_json",
+                });
+            }
+
+            const result = await withAudioFallback(c, async (candidate) => {
+                if (candidate.id === "grok-transcribe") {
+                    return transcribeWithXai({
+                        file,
+                        language: language || undefined,
+                        responseFormat: responseFormat || undefined,
+                        apiKey: c.env.XAI_API_KEY,
+                        log,
+                    });
+                }
+
+                if (candidate.id === "scribe") {
+                    return transcribeWithElevenLabs({
+                        file,
+                        language: language || undefined,
+                        responseFormat: responseFormat || undefined,
+                        apiKey: c.env.ELEVENLABS_API_KEY,
+                        log,
+                        numSpeakers: speakersExpected,
+                    });
+                }
+
+                if (
+                    candidate.id === "universal-2" ||
+                    candidate.id === "universal-3.5-pro"
+                ) {
+                    return transcribeWithAssemblyAi({
+                        file,
+                        language: language || undefined,
+                        prompt: prompt || undefined,
+                        responseFormat: responseFormat || undefined,
+                        temperature,
+                        model: candidate.id,
+                        apiKey: (
+                            c.env as unknown as {
+                                ASSEMBLYAI_API_KEY: string;
+                            }
+                        ).ASSEMBLYAI_API_KEY,
+                        log,
+                        speakersExpected,
+                    });
+                }
+
+                const ovhApiKey = c.env.OVHCLOUD_API_KEY;
+                if (!ovhApiKey) {
+                    throw new UpstreamError(500 as ContentfulStatusCode, {
+                        message:
+                            "Transcription service is not configured (missing API key)",
+                    });
+                }
+                validateWhisperResponseFormat(responseFormat);
+
+                // Rebuild the consumed multipart body for each upstream attempt.
+                const whisperFormData = new FormData();
+                const filename =
+                    file.name && file.name !== "blob" ? file.name : "audio.mp3";
+                whisperFormData.append("file", file, filename);
+                if (language) whisperFormData.append("language", language);
+                whisperFormData.append("response_format", "verbose_json");
+                whisperFormData.append("model", "whisper-large-v3");
+                whisperFormData.append("timestamp_granularities[]", "word");
+
+                const whisperUrl =
+                    "https://oai.endpoints.kepler.ai.cloud.ovh.net/v1/audio/transcriptions";
+                const response = await ensureUpstreamOk(
+                    await fetch(whisperUrl, {
+                        method: "POST",
+                        headers: { Authorization: `Bearer ${ovhApiKey}` },
+                        body: whisperFormData,
+                    }),
+                    whisperUrl,
+                );
+
+                let whisper: WhisperVerboseJson;
+                try {
+                    whisper = JSON.parse(await response.text());
+                } catch {
+                    throw new UpstreamError(502 as ContentfulStatusCode, {
+                        message:
+                            "Whisper returned an unexpected (non-JSON) response",
+                    });
+                }
+                const usageHeaders = buildUsageHeaders(
+                    candidate.id,
+                    createAudioSecondsUsage(extractWhisperUsage(whisper, log)),
+                );
+                return formatWhisperResponse(
+                    whisper,
+                    responseFormat,
+                    usageHeaders,
+                );
+            });
+            c.var.track.overrideResponseTracking(result.clone());
+            return result;
+        },
     );
 
 export function parsePositiveInt(
