+++ gen.pollinations.ai/src/routes/images.ts
@@ -30,6 +30,7 @@ import type { Env } from "@/env.ts";
 import { generateImageOrVideoResponse } from "@/image/handler.ts";
 import { applySafety, withSafetyHeaders } from "@/middleware/safety.ts";
 import { arrayBufferToBase64 } from "@/util.ts";
+import { requireGenerationAccess } from "@/utils/generation-access.ts";
 
 // --- Helpers ---
 
@@ -246,7 +247,6 @@ export const prepareOpenAIImageGeneration = createMiddleware<Env>(
             body.safe as SafeValue,
         );
         Object.assign(body, resolved, { model, prompt: safePrompt });
-        c.set("generationRequestBody", JSON.stringify(body));
 
         const imageUrl = new URL(
             `/image/${encodeURIComponent(body.prompt)}`,
@@ -336,6 +336,8 @@ export async function handleImageGeneration(c: Context<Env>) {
 }
 
 export async function handleImageEdit(c: Context<Env>) {
+    await requireGenerationAccess(c.var, c.env);
+
     const { prompt, imageUrls, size, quality, seed, safe, extra } =
         await parseEditInput(c);
     const safePrompt = await applySafety(c, prompt, safe);
