+++ gen.pollinations.ai/src/routes/proxy.ts
@@ -1,12 +1,14 @@
 import { type Context, Hono } from "hono";
-import { every } from "hono/combine";
+import { bodyLimit } from "hono/body-limit";
 import { resolver as baseResolver, describeRoute } from "hono-openapi";
+import {
+    generateEmbeddings,
+    getEmbeddingProviderModelId,
+} from "@/embeddings/handler.ts";
 import type { Env } from "@/env.ts";
-import { handleRegisterServer } from "@/image/handler.ts";
+import { handleImagePrompt, handleRegisterServer } from "@/image/handler.ts";
 import { auth } from "@/middleware/auth.ts";
 import { balance } from "@/middleware/balance.ts";
-import { prepareGenerationRequest } from "@/middleware/generation-cache.ts";
-import { deduplicateGeneration } from "@/middleware/generation-deduplication.ts";
 import {
     audioCache,
     imageCache,
@@ -29,6 +31,7 @@ import {
 const resolver = <T extends Parameters<typeof baseResolver>[0]>(schema: T) =>
     baseResolver(schema, { reused: "ref" });
 
+import { UpstreamError } from "@shared/error.ts";
 import { validator } from "@shared/middleware/validator.ts";
 import { AUDIO_VOICES } from "@shared/registry/audio.ts";
 import {
@@ -45,16 +48,25 @@ import {
     REALTIME_MODEL_NAMES,
 } from "@shared/registry/realtime.ts";
 import {
+    type CreateChatCompletionRequest,
     CreateChatCompletionRequestSchema,
+    type CreateChatCompletionResponse,
     CreateChatCompletionResponseSchema,
     CreateImageRequestSchema,
     CreateImageResponseSchema,
     GetModelsResponseSchema,
 } from "@shared/schemas/openai.ts";
-import { SafeSchema } from "@shared/schemas/safety.ts";
+import { SafeSchema, type SafeValue } from "@shared/schemas/safety.ts";
 import { errorResponseDescriptions } from "@shared/utils/api-docs.ts";
 import { createFactory } from "hono/factory";
 import { z } from "zod";
+import {
+    applySafety,
+    applySafetyToChatRequest,
+    applySafetyToTexts,
+    withSafetyHeaders,
+} from "@/middleware/safety.ts";
+import { handle3dPrompt } from "@/model3d/handler.ts";
 import {
     CreateEmbeddingRequestSchema,
     CreateEmbeddingResponseSchema,
@@ -69,22 +81,18 @@ import {
 } from "@/schemas/model3d.ts";
 import { RealtimeRequestQueryParamsSchema } from "@/schemas/realtime.ts";
 import { GenerateTextRequestQueryParamsSchema } from "@/schemas/text.ts";
+import {
+    handleChatCompletionLocal,
+    handleSimpleTextLocal,
+    handleTextContentLocal,
+} from "@/text/handler.ts";
 import { generationAccess } from "@/utils/generation-access.ts";
+import { withModelFallbackResponse } from "../fallback.ts";
 import {
     type GenerationModelEntry,
     getGenerationModelRegistry,
 } from "../model-registry.ts";
 import { handleSimpleAudio } from "./audio.ts";
-import {
-    generateChatCompletion,
-    generateEmbeddingsResponse,
-    generateImageVideo,
-    generateModel3d,
-    generateSimpleText,
-    generateTextContent,
-    simpleAudioQuerySchema,
-    textBodyLimit,
-} from "./generation-handlers.ts";
 import { handleRealtimeWebSocket } from "./realtime.ts";
 
 // Build dynamic model lists from registry for use in API descriptions
@@ -133,26 +141,44 @@ function describeRealtimeWebSocket(path: "/realtime" | "/v1/realtime") {
 }
 
 const factory = createFactory<Env>();
-
+const textBodyLimit = bodyLimit({
+    maxSize: 20 * 1024 * 1024,
+});
 // Shared handler for image and video generation (used by both /image/ and /video/ routes)
 const imageVideoHandlers = factory.createHandlers(
     track("generate.image"),
     imageCache,
     generationAccess,
-    deduplicateGeneration,
-    generateImageVideo,
+    async (c) => {
+        const query = c.req.valid("query" as never) as { safe?: SafeValue };
+        const prompt = await applySafety(
+            c,
+            c.req.param("prompt") || "",
+            query.safe,
+        );
+        return withSafetyHeaders(c, await handleImagePrompt(c, prompt));
+    },
 );
 
-// 3D shares the image event type to avoid changing Tinybird consumers.
+// Handler for 3D model generation (reuses the "generate.image" EventType,
+// same as video, to avoid touching Tinybird/EventType consumers).
 const model3dHandlers = factory.createHandlers(
     resolveModel("generate.image", { defaultModel: DEFAULT_3D_MODEL }),
     track("generate.image"),
     model3dCache,
     generationAccess,
-    deduplicateGeneration,
-    generateModel3d,
+    async (c) => {
+        const query = c.req.valid("query" as never) as { safe?: SafeValue };
+        const prompt = await applySafety(
+            c,
+            c.req.param("prompt") || "",
+            query.safe,
+        );
+        return withSafetyHeaders(c, await handle3dPrompt(c, prompt));
+    },
 );
 
+// Shared handler for OpenAI-compatible chat completions
 const chatCompletionHandlers = factory.createHandlers(
     textBodyLimit,
     validator("json", CreateChatCompletionRequestSchema),
@@ -160,9 +186,69 @@ const chatCompletionHandlers = factory.createHandlers(
     track("generate.text"),
     textCache,
     generationAccess,
-    deduplicateGeneration,
-    generateChatCompletion,
+    async (c) => {
+        // Use resolved model from middleware for the backend request
+        const requestBody = await applySafetyToChatRequest(c, {
+            ...(c.req.valid("json" as never) as CreateChatCompletionRequest),
+            model: c.var.model.resolved,
+        });
+
+        const response = await handleChatCompletionLocal(c, requestBody);
+        if (!response.ok) return response;
+
+        assertStreamContentType(c, response, c.var.upstreamRequestUrl);
+
+        // add content filter headers if not streaming
+        let contentFilterHeaders = {};
+        if (!c.var.track.streamRequested) {
+            const responseText = await response.clone().text();
+            try {
+                const parsedResponse = CreateChatCompletionResponseSchema.parse(
+                    JSON.parse(responseText),
+                    { reportInput: true },
+                );
+                contentFilterHeaders =
+                    contentFilterResultsToHeaders(parsedResponse);
+            } catch (parseError) {
+                throw new UpstreamError(502, {
+                    message: `Upstream returned response that failed schema validation: ${parseError instanceof Error ? parseError.message : String(parseError)}`,
+                    requestUrl: c.var.upstreamRequestUrl,
+                    responseBody: responseText,
+                    cause: parseError,
+                });
+            }
+        }
+
+        return withSafetyHeaders(
+            c,
+            new Response(response.body, {
+                headers: {
+                    ...Object.fromEntries(response.headers),
+                    ...contentFilterHeaders,
+                },
+            }),
+        );
+    },
 );
+
+// Validate streaming responses: if client requested stream but upstream
+// returned non-SSE, throw rather than forwarding broken data.
+function assertStreamContentType(
+    c: Context<Env>,
+    response: Response,
+    upstreamRequestUrl: URL | undefined,
+): void {
+    if (c.var.track.streamRequested) {
+        const contentType = response.headers.get("content-type") || "";
+        if (!contentType.includes("text/event-stream")) {
+            throw new UpstreamError(502, {
+                message: `Stream requested for model ${c.var.model.resolved} but upstream returned content-type: ${contentType}`,
+                requestUrl: upstreamRequestUrl,
+                responseBody: contentType,
+            });
+        }
+    }
+}
 
 // Helper to filter models by API key permissions and paid balance.
 function filterEntriesByPermissions(
@@ -582,11 +668,28 @@ export const proxyRoutes = new Hono<Env>()
         validator("json", CreateEmbeddingRequestSchema),
         resolveModel("generate.embedding"),
         track("generate.embedding"),
-        prepareGenerationRequest,
-        textCache,
         generationAccess,
-        deduplicateGeneration,
-        generateEmbeddingsResponse,
+        async (c) => {
+            const requestBody = c.req.valid("json" as never) as z.infer<
+                typeof CreateEmbeddingRequestSchema
+            >;
+            const { response, servedEntry } = await withModelFallbackResponse(
+                c.var.model,
+                (candidate) =>
+                    generateEmbeddings(
+                        c.env,
+                        {
+                            ...requestBody,
+                            model: getEmbeddingProviderModelId(candidate.id),
+                        },
+                        candidate.definition ?? c.var.model.definition,
+                        candidate.id,
+                    ),
+                c.var.track?.failedCalls,
+            );
+            if (servedEntry) c.set("servedModelEntry", servedEntry);
+            return response;
+        },
     )
     .post(
         "/text",
@@ -612,8 +715,18 @@ export const proxyRoutes = new Hono<Env>()
         track("generate.text"),
         textCache,
         generationAccess,
-        deduplicateGeneration,
-        generateTextContent,
+        async (c) => {
+            const requestBody = await applySafetyToChatRequest(c, {
+                ...(c.req.valid(
+                    "json" as never,
+                ) as CreateChatCompletionRequest),
+                model: c.var.model.resolved,
+            });
+
+            const response = await handleTextContentLocal(c, requestBody);
+            assertStreamContentType(c, response, c.var.upstreamRequestUrl);
+            return withSafetyHeaders(c, response);
+        },
     )
     .get(
         "/text/:prompt{[\\s\\S]+}",
@@ -651,8 +764,34 @@ export const proxyRoutes = new Hono<Env>()
         track("generate.text"),
         textCache,
         generationAccess,
-        deduplicateGeneration,
-        generateSimpleText,
+        async (c) => {
+            // Use resolved model from middleware
+            const model = c.var.model.resolved;
+
+            const query = c.req.valid("query" as never) as {
+                safe?: SafeValue;
+                system?: string;
+            };
+            const textInputs =
+                typeof query.system === "string"
+                    ? [c.req.param("prompt"), query.system]
+                    : [c.req.param("prompt")];
+            const [prompt, system] = await applySafetyToTexts(
+                c,
+                textInputs,
+                query.safe,
+            );
+
+            return withSafetyHeaders(
+                c,
+                await handleSimpleTextLocal(
+                    c,
+                    prompt,
+                    model,
+                    system ? { system } : undefined,
+                ),
+            );
+        },
     )
     .get(
         // Use :prompt{[\\s\\S]+} regex to capture everything including slashes AND newlines
@@ -801,7 +940,7 @@ export const proxyRoutes = new Hono<Env>()
             tags: ["🧊 3D"],
             summary: "Generate 3D Model With JSON",
             description:
-                "Generate a 3D model from a text prompt or reference image using JSON parameters. `trellis-2` supports `low`, `medium`, and `high` resolution with variable pricing.",
+                "Generate a 3D model from a text prompt or reference image using JSON parameters. POST requests bypass the server-side media cache, so repeated requests regenerate and are billed. `trellis-2` supports `low`, `medium`, and `high` resolution with variable pricing.",
             responses: {
                 200: {
                     description: "Success - Returns the generated 3D model",
@@ -842,10 +981,18 @@ export const proxyRoutes = new Hono<Env>()
         validator("json", Generate3dRequestBodySchema),
         resolveModel("generate.image", { defaultModel: DEFAULT_3D_MODEL }),
         track("generate.image"),
-        every(prepareGenerationRequest, model3dCache),
         generationAccess,
-        deduplicateGeneration,
-        generateModel3d,
+        async (c) => {
+            const query = c.req.valid("query" as never) as {
+                safe?: SafeValue;
+            };
+            const prompt = await applySafety(
+                c,
+                c.req.param("prompt") || "",
+                query.safe,
+            );
+            return withSafetyHeaders(c, await handle3dPrompt(c, prompt));
+        },
     )
     .get(
         "/audio/:text",
@@ -885,14 +1032,113 @@ export const proxyRoutes = new Hono<Env>()
                 }),
             }),
         ),
-        validator("query", simpleAudioQuerySchema),
+        validator(
+            "query",
+            z.object({
+                voice: z
+                    .enum(AUDIO_VOICES as unknown as [string, ...string[]])
+                    .default("alloy")
+                    .meta({
+                        description:
+                            "Voice to use for speech generation (TTS only)",
+                        example: "nova",
+                    }),
+                response_format: z
+                    .enum(["mp3", "opus", "aac", "flac", "wav", "pcm"])
+                    .default("mp3")
+                    .meta({
+                        description:
+                            "Audio output format. CSM and Kokoro support mp3, opus, flac, wav, and pcm; Qwen TTS currently returns WAV regardless of this setting; lyria-3-clip and eleven-sfx support mp3 only.",
+                        example: "mp3",
+                    }),
+                model: z.string().optional().meta({
+                    description:
+                        "Audio model: TTS (default) or a music-generation model such as lyria-3-clip",
+                    example: "tts-1",
+                }),
+                duration: z
+                    .string()
+                    .optional()
+                    .transform((v) => (v ? parseFloat(v) : undefined))
+                    .pipe(z.number().min(0.5).max(300).optional())
+                    .meta({
+                        description:
+                            "Music duration in seconds (elevenmusic 3-300; lyria-3-clip fixed at 30)",
+                        example: "30",
+                    }),
+                seconds: z.coerce.number().min(1).max(380).optional().meta({
+                    description:
+                        "Audio duration in seconds for stable-audio-3-medium/large, 1-380",
+                    example: "30",
+                }),
+                steps: z.coerce.number().int().min(1).max(100).optional().meta({
+                    description:
+                        "Sampling steps (stable-audio-3-medium 1-100, stable-audio-3-large 4-8)",
+                    example: "8",
+                }),
+                negative_prompt: z.string().optional().meta({
+                    description: "Negative prompt for stable-audio-3-large",
+                    example: "distortion, vocals",
+                }),
+                instrumental: z
+                    .enum(["true", "false"])
+                    .default("false")
+                    .transform((v) => v === "true")
+                    .meta({
+                        description:
+                            "If true, guarantees instrumental output (elevenmusic only)",
+                        example: "false",
+                    }),
+                instruct: z.string().optional().meta({
+                    description:
+                        "Emotion/style instruction (qwen-tts-instruct only)",
+                    example: "speak softly and warmly",
+                }),
+                loop: z
+                    .enum(["true", "false"])
+                    .optional()
+                    .transform((v) =>
+                        v === undefined ? undefined : v === "true",
+                    )
+                    .meta({
+                        description:
+                            "Loop the generated sound effect (eleven-sfx only)",
+                        example: "false",
+                    }),
+                prompt_influence: z
+                    .string()
+                    .optional()
+                    .transform((v) => (v ? Number.parseFloat(v) : undefined))
+                    .pipe(z.number().min(0).max(1).optional())
+                    .meta({
+                        description:
+                            "How strictly to follow the prompt, 0-1 (eleven-sfx only)",
+                        example: "0.3",
+                    }),
+                seed: z.coerce
+                    .number()
+                    .int()
+                    .min(-1)
+                    .max(4294967295)
+                    .optional()
+                    .meta({
+                        description:
+                            "Seed passed to the model. Same seed + parameters return the same cached result while available.",
+                        example: "42",
+                    }),
+                key: z.string().optional().meta({
+                    description:
+                        "API key (alternative to Authorization header)",
+                }),
+                safe: SafeSchema,
+            }),
+        ),
         resolveModel("generate.audio", {
             supportedEndpoint: "/audio/{text}",
         }),
         track("generate.audio"),
         audioCache,
         generationAccess,
-        deduplicateGeneration,
         handleSimpleAudio,
     )
     .post(
@@ -923,10 +1169,9 @@ export const proxyRoutes = new Hono<Env>()
         resolveModel("generate.image"),
         track("generate.image"),
         generationAccess,
-        every(prepareOpenAIImageGeneration, formatOpenAIImageGeneration),
-        prepareGenerationRequest,
+        prepareOpenAIImageGeneration,
+        formatOpenAIImageGeneration,
         imageCache,
-        deduplicateGeneration,
         handleImageGeneration,
     )
     .post(
@@ -957,9 +1202,74 @@ export const proxyRoutes = new Hono<Env>()
         }),
         resolveModel("generate.image", { defaultModel: "flux" }),
         track("generate.image"),
-        prepareGenerationRequest,
-        textCache,
-        generationAccess,
-        deduplicateGeneration,
         handleImageEdit,
     );
+
+export function contentFilterResultsToHeaders(
+    response: CreateChatCompletionResponse,
+): Record<string, string> {
+    const promptFilters =
+        response.prompt_filter_results?.[0]?.content_filter_results;
+    const completionFilters = response.choices?.[0]?.content_filter_results;
+
+    const mapToString = (value: unknown): string | undefined =>
+        value ? String(value) : undefined;
+
+    // Build header mappings
+    const headerMappings: Array<[string, unknown]> = [
+        // Prompt filters
+        ["x-moderation-prompt-hate-severity", promptFilters?.hate?.severity],
+        [
+            "x-moderation-prompt-self-harm-severity",
+            promptFilters?.self_harm?.severity,
+        ],
+        [
+            "x-moderation-prompt-sexual-severity",
+            promptFilters?.sexual?.severity,
+        ],
+        [
+            "x-moderation-prompt-violence-severity",
+            promptFilters?.violence?.severity,
+        ],
+        [
+            "x-moderation-prompt-jailbreak-detected",
+            promptFilters?.jailbreak?.detected,
+        ],
+        // Completion filters
+        [
+            "x-moderation-completion-hate-severity",
+            completionFilters?.hate?.severity,
+        ],
+        [
+            "x-moderation-completion-self-harm-severity",
+            completionFilters?.self_harm?.severity,
+        ],
+        [
+            "x-moderation-completion-sexual-severity",
+            completionFilters?.sexual?.severity,
+        ],
+        [
+            "x-moderation-completion-violence-severity",
+            completionFilters?.violence?.severity,
+        ],
+        [
+            "x-moderation-completion-protected-material-text-detected",
+            completionFilters?.protected_material_text?.detected,
+        ],
+        [
+            "x-moderation-completion-protected-material-code-detected",
+            completionFilters?.protected_material_code?.detected,
+        ],
+    ];
+
+    // Convert to headers, filtering out undefined values
+    const headers: Record<string, string> = {};
+    for (const [key, value] of headerMappings) {
+        const stringValue = mapToString(value);
+        if (stringValue !== undefined) {
+            headers[key] = stringValue;
+        }
+    }
+
+    return headers;
+}
