--- gen.pollinations.ai/src/utils/execute-generation.ts
@@ -1,133 +1,0 @@
-import type { BalanceCheckResult } from "@shared/billing/balance.ts";
-import { handleError } from "@shared/error.ts";
-import { Hono } from "hono";
-import type { Env } from "@/env.ts";
-import {
-    authFromSnapshot,
-    type GenerationAuthSnapshot,
-} from "@/middleware/auth.ts";
-import { balance } from "@/middleware/balance.ts";
-import type {
-    GenerationErrorSnapshot,
-    GenerationOutcome,
-} from "@/middleware/generation-deduplication.ts";
-import { logger } from "@/middleware/logger.ts";
-import { frontendKeyBilling } from "@/middleware/rate-limit-durable.ts";
-import { generationExecutorRoutes } from "@/routes/generation-executor.ts";
-
-async function drainResponse(response: Response): Promise<void> {
-    const reader = response.body?.getReader();
-    if (!reader) return;
-    while (!(await reader.read()).done) {
-        // Draining completes the existing cache transform/clone.
-    }
-}
-
-const ERROR_BODY_BYTES = 64 * 1024;
-const ERROR_HEADERS = new Set(["content-type", "retry-after"]);
-const ERROR_HEADER_PREFIXES = ["x-moderation-", "x-safety-"];
-
-async function captureError(
-    response: Response,
-): Promise<GenerationErrorSnapshot> {
-    const body = new Uint8Array(
-        (await response.arrayBuffer()).slice(0, ERROR_BODY_BYTES),
-    );
-    return {
-        httpStatus: response.status,
-        headers: [...response.headers.entries()].filter(([name]) => {
-            const lowerName = name.toLowerCase();
-            return (
-                ERROR_HEADERS.has(lowerName) ||
-                ERROR_HEADER_PREFIXES.some((prefix) =>
-                    lowerName.startsWith(prefix),
-                )
-            );
-        }),
-        body,
-    };
-}
-
-export type DetachedGeneration = {
-    result: GenerationOutcome;
-    settlement: Promise<void>;
-};
-
-function generationExecutor(
-    auth: GenerationAuthSnapshot,
-    requestId: string,
-    balanceCheckResult: BalanceCheckResult,
-    registerGenerationCacheWrite: (promise: Promise<void>) => void,
-): Hono<Env> {
-    const executor = new Hono<Env>()
-        .use("*", async (c, next) => {
-            c.set("requestId", requestId);
-            c.header("X-Request-Id", requestId);
-            await next();
-        })
-        .use("*", logger)
-        .use("*", authFromSnapshot(auth))
-        .use("*", async (c, next) => {
-            c.set("registerGenerationCacheWrite", registerGenerationCacheWrite);
-            await next();
-        })
-        .use("*", frontendKeyBilling)
-        .use("*", balance)
-        .use("*", async (c, next) => {
-            c.var.balance.balanceCheckResult = balanceCheckResult;
-            await next();
-        })
-        .route("/", generationExecutorRoutes);
-    executor.onError(handleError);
-    return executor;
-}
-
-/** Runs a provider handler under the Durable Object alarm's lifetime. */
-export async function executeGeneration(
-    request: Request,
-    auth: GenerationAuthSnapshot,
-    requestId: string,
-    balanceCheckResult: BalanceCheckResult,
-    env: CloudflareBindings,
-): Promise<DetachedGeneration> {
-    const promises: Promise<unknown>[] = [];
-    let cacheWrite: Promise<void> | undefined;
-    const executionCtx = {
-        waitUntil(promise: Promise<unknown>) {
-            promises.push(promise);
-        },
-        passThroughOnException() {},
-    } as ExecutionContext;
-
-    const response = await generationExecutor(
-        auth,
-        requestId,
-        balanceCheckResult,
-        (promise) => {
-            cacheWrite = promise;
-        },
-    ).fetch(request, env, executionCtx);
-    const settlement = Promise.allSettled(promises).then(() => {});
-
-    try {
-        const failed = !response.ok;
-        const error = failed ? await captureError(response) : undefined;
-        if (!failed) await drainResponse(response);
-
-        if (response.headers.get("x-cache") === "HIT") {
-            return { result: { status: "cached" }, settlement };
-        }
-        if (error) {
-            return { result: { status: "failed", error }, settlement };
-        }
-        if (!cacheWrite) {
-            throw new Error("Generation completed without a cacheable result");
-        }
-
-        await cacheWrite;
-        return { result: { status: "cached" }, settlement };
-    } catch (error) {
-        await settlement;
-        throw error;
-    }
-}
