+++ gen.pollinations.ai/test/elevenlabs-voice-isolator.test.ts
@@ -1,9 +1,7 @@
-import { createExecutionContext, env } from "cloudflare:test";
+import { env, SELF } from "cloudflare:test";
 import { test as workerTest } from "@shared/test/fixtures/index.ts";
 import { afterEach, describe, expect, it, vi } from "vitest";
-import worker from "../src/index.ts";
 import { isolateVoiceWithElevenLabs } from "../src/routes/audio.ts";
-import { withInlineGenerationCoordinator } from "./helpers/inline-generation-coordinator.ts";
 
 const errorLog = vi.fn();
 const log = {
@@ -11,14 +9,6 @@ const log = {
     info: vi.fn(),
     warn: vi.fn(),
 } as never;
-
-async function fetchGen(input: RequestInfo | URL, init?: RequestInit) {
-    return worker.fetch(
-        new Request(input, init),
-        withInlineGenerationCoordinator(env),
-        createExecutionContext(),
-    );
-}
 
 function createWav(seconds: number): File {
     const sampleRate = 16000;
@@ -125,7 +115,7 @@ workerTest(
         formData.append("model", "eleven-voice-isolator");
         formData.append("input", "Not an isolation request.");
 
-        const response = await fetchGen(
+        const response = await SELF.fetch(
             "https://gen.pollinations.ai/v1/audio/speech",
             {
                 method: "POST",
@@ -149,7 +139,7 @@ workerTest.runIf(Boolean(env.ELEVENLABS_API_KEY))(
         formData.append("model", "voice-isolator");
         formData.append("audio", createWav(5));
 
-        const response = await fetchGen(
+        const response = await SELF.fetch(
             "https://gen.pollinations.ai/v1/audio/voice-isolator",
             {
                 method: "POST",
