--- gen.pollinations.ai/test/helpers/inline-generation-coordinator.ts
@@ -1,31 +1,0 @@
-import type { GenerationJob } from "@/middleware/generation-deduplication.ts";
-import { executeGeneration } from "@/utils/execute-generation.ts";
-
-/** Keeps provider mocks in one isolate while exercising the real middleware. */
-export function withInlineGenerationCoordinator(
-    bindings: CloudflareBindings,
-): CloudflareBindings {
-    const coordinated = { ...bindings } as CloudflareBindings;
-    coordinated.GENERATION_COORDINATOR = {
-        getByName() {
-            return {
-                async startAndWait(job: GenerationJob) {
-                    const execution = await executeGeneration(
-                        new Request(job.request.url, {
-                            method: job.request.method,
-                            headers: job.request.headers,
-                            body: job.request.body?.slice().buffer,
-                        }),
-                        job.auth,
-                        job.requestId,
-                        job.balanceCheckResult,
-                        coordinated,
-                    );
-                    await execution.settlement;
-                    return execution.result;
-                },
-            } as never;
-        },
-    } as never;
-    return coordinated;
-}
