+++ gen.pollinations.ai/vitest.config.ts
@@ -14,7 +14,6 @@ const genAliases = [
     "content-filter.ts",
     "cache",
     "durable-objects/PollenRateLimiter.ts",
-    "durable-objects/GenerationCoordinator.ts",
     "env.ts",
     "error.ts",
     "events.ts",
@@ -22,8 +21,6 @@ const genAliases = [
     "logger.ts",
     "middleware/auth.ts",
     "middleware/balance.ts",
-    "middleware/generation-cache.ts",
-    "middleware/generation-deduplication.ts",
     "middleware/logger.ts",
     "middleware/media-cache.ts",
     "middleware/model.ts",
@@ -33,7 +30,6 @@ const genAliases = [
     "middleware/text-cache.ts",
     "middleware/track.ts",
     "middleware/validator.ts",
-    "routes/generation-executor.ts",
     "schemas/embeddings.ts",
     "schemas/image.ts",
     "schemas/model3d.ts",
@@ -44,7 +40,6 @@ const genAliases = [
     "util.ts",
     "utils/api-docs.ts",
     "utils/bedrock-guardrail.ts",
-    "utils/execute-generation.ts",
     "utils/generation-access.ts",
     "utils/media-cache.ts",
     "utils/model-stats.ts",
@@ -54,7 +49,6 @@ const genAliases = [
 
 const baseConfig = defineConfig({
     resolve: {
-        dedupe: ["hono", "hono-openapi"],
         alias: [
             ...genAliases.map((path) => ({
                 find: `@/${path}`,
